Create BpkIconChip
