+++ packages/bpk-component-chip/src/BpkSelectableChip.tsx
@@ -37,6 +37,7 @@ const BpkSelectableChip = ({
   children,
   className,
   disabled = false,
+  leadingAccessoryClassName,
   leadingAccessoryView = null,
   role = 'checkbox',
   selected = false,
@@ -64,7 +65,7 @@ const BpkSelectableChip = ({
       {...rest}
     >
       {leadingAccessoryView && (
-        <span className={getClassName('bpk-chip__leading-accessory-view')}>
+        <span className={getClassName('bpk-chip__leading-accessory-view', leadingAccessoryClassName)}>
           {leadingAccessoryView}
         </span>
       )}
