+++ packages/bpk-mixins/_chips.scss
@@ -237,4 +237,13 @@
     color: $bpk-text-disabled-day;
     cursor: not-allowed;
   }
+
+  &--icon-only {
+    padding-inline-end: bpk-spacing-md();
+    padding-inline-start: bpk-spacing-md();
+
+    &__leading-accessory-view {
+      margin-inline-end: 0;
+    }
+  }
 }
